backend/requirements.txt
httpx~=0.23.0
fastapi~=0.85.0
beautifulsoup4~=4.11.1
lxml~=4.9.1
uvicorn~=0.18.3
CaseInsensitiveDict~=1.0.0
tortoise-orm~=0.19.2
aiopg~=1.3.5
orjson~=3.8.0
rich~=12.6.0
aiofiles~=0.8.0
dpath~=2.0.6